requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-environ==0.10.0
django-filter==23.2
django-mongoengine==0.5.4
django-rest-framework==0.1.0
django-rest-framework-mongoengine==3.4.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
docker==6.1.3
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mongoengine==0.27.0
mysql-client==0.0.1
openapi-codec==1.3.2
packaging==23.1
psycopg2-binary==2.9.6
PyJWT==2.7.0
pymongo==4.4.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
simplejson==3.19.1
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
websocket-client==1.6.0
PyMySQL==1.0.3